requirements.txt
accelerate==0.16.0
transformers==4.26.1
ftfy==6.1.1
Pillow==9.2.0
diffusers==0.14.0
git+https://github.com/cloneofsimo/lora.git@799c17aef2a475641fb70d68a6992de4fc325ce4
xformers
clip-interrogator==0.5.4
matplotlib
fastapi==0.89.1
uvicorn[standard]==0.20.0
opencv-contrib-python==4.7.0.72